refactor(options): nest the site rules folder under translation control
The page moved to /page-translation/translation-control/site-rules but its
source stayed at pages/site-rules — a top-level folder for a page three
levels deep in the UI. Every other detail page already sits inside the
section that drills into it; this one now does too.
