+++ src/entrypoints/options/pages/translation/translation-control/site-rules/__tests__/built-in-rules.test.tsx
