+++ src/entrypoints/options/pages/translation/translation-control/site-rules/__tests__/user-rules-editor.test.tsx
