+++ src/entrypoints/options/pages/translation/translation-control/site-rules/__tests__/validate-user-rules.test.ts
