+++ src/entrypoints/options/pages/translation/translation-control/site-rules/built-in-rule-row.tsx
