+++ src/entrypoints/options/pages/translation/translation-control/site-rules/validate-user-rules.ts
